Increase timeout of STWClient
Timeout was increased by building an OkHttpClient with a timeout of 60
seconds for connect, read, and write.
The timeout needed to be increased because showstowatch.josephtran.io
is using heroku's free plan and therefore falls asleep after 30 mins of
inactivity.

Update build.gradle to use okhttp, so that OkHttpClients could be
created