Add login fragment and presenter to login to STW
Update STWClient to retrieve the login and redirect url, this is in the client,
instead of STWService, in case the STW api creates these values based on user.
Move STW url from STWClient to STWService.

Add factory method to create HomeFragment.

Update MainActivity to load a LoginFragment on start and HomeFragment on
log in.