Update STW api to include STW add, delete, update, and search actions
Make STWClient not return observables since the client should only be
responsible for interacting with the api.
Add STWClientWrapper to wrap STWClient's functions with observables.
It is also responsible for retrieving the user's name and token from the shared
preferences.
Update STWShow to include data for a show's id and tvdb id.
Add TVDBShow to represent a TVDB show when searching for a show's tvdb id.

Update presenters to use STWClientWrapper.
Add user preference key to MainActivity.
Update LoginPresenter to save the user's email when logging in.

Update gradle build tool to 2.2.3.