Refactor home feature to a more general shows feature
Shows will be used to represent more than just all shows from STW's api
(e.g. in progress, completed, etc.).
Update STWShow to include abandoned and completed values.
Change the layout used to display shows to display shows in a grid rather
than a list.
Remove season and episode info from a show item in the adapter.
Add an overlay to represent the status of the show.
Decrease the size of the progress bar in the shows fragment.

Add dark red and dark green colors to represent abandoned and
completed shows.