Add a show form view
Add ShowFormFragment to inflate the show form view.
It currently does nothing more than inflate the view.
Add ShowFormActivity to contain the ShowFormFragment.
Add a fab to ShowsActivity to start the ShowFormActivity.

Increase min sdk version to 21 so that button and background
tints could be used without having to create different layouts for lower
versions.