Handle unauthorized user actions
STWClient now throws an STWUnauthorizedException if the current user could not
be authorized.

Update show add classes to handle when this exception is thrown.
An error message is displayed and the show is not added to STW.
Add showErrorMessage to ShowAddContracts - to show error messages in red.
Fix bug with showMessages not using white text color.
Fix bug with showMessages not validating the correct title view.

Add ShowsPagerAdapter to represent the anonymous FragmentPagerAdapter
previously defined in ShowsActivity.
Update ShowsActivity to use ShowsPagerAdapter.
Modify ShowsActivity to display a success message and refresh the pager after
the user has added a show.
Modify ShowsFragment to always clear the adapter if it is not empty, so that
pages can be refreshed properly.