Add TVDB search when adding shows to set a show's TVDB id/banner
Make banner and tvdbId in STWShow's mutable since they are now updated
via show search.

Modify ShowAddActivity to handle when shows have been added or searched.
Add interface to interact with the activity when a show has been added in ShowAddFragment.
Rename ShowFormView to ShowAddView since it is only used for adding shows.
Modify ShowAddFormView's closeView to use onShowAdded to better reflect its usage.

Modify ShowsActivity to always refresh the adapter after the shows activity has been closed.
This ensures the pager is up to date even if the user didn't select a tvdb show.
Change the visibility of STWShowsAdapter's shows to private and modify ShowsFragment
to clear the adapter instead of directly clearing the list.

Add id to the relative layout containing the toolbar button in activity_show_form, so
that it could be removed dynamically.
Remove tools:context from fragment_show_form since it may be used with other fragments.

Rename stw_show_item.xml to show_item since it is now used with tvdb shows
when searching as well.
Adjusted alpha to 0.5 for its overlay.

Add missing drawable used for shows activity's fab