Fix bug with deleting shows crashing when using STWClientWrapper
STWClient's deleteShow would fail to create a JSONObjectConverter when
using STWClientWrapper (i.e. Observables). deleteShow now just attempts to
parse for a String instead of a JSONObject.

Fix bug with STWShow not storing the tvdb id when retrieving it from STW.

Modify ShowSearchPresenter to set the current show's title to the tvdb show
when setting the TVDB show as well.