Redesign the show form to include an image view and use lighter colors
Add a collapsing toolbar to the show form activity. It holds an image
and title for subclasses to set - as well as the existing toolbar.
Add two abstract functions to ShowFormActivity to allow subclasses to
load their own images and title to the show form toolbar.
Implement these two functions in ShowAddActivity and ShowEditActivity.
When adding, show_add_toolbar_bg is displayed in the toolbar.
When editing, the show's banner is now displayed in the toolbar.

Change to use lighter colors in the show form fragment.

Update to version 1.0.6 of Kotlin.